MAINT: optimize: separate `_constraints` and `_canonical_constraint`
The basic idea is that `_constraints` contains "public" classes
while `_canonical_constraint` contains "private" classes and
functions, reserved for internal use.
