+++ scipy/optimize/_large_scale_constrained/__init__.py
@@ -2,6 +2,7 @@
 
 
 from .equality_constrained_sqp import equality_constrained_sqp
+from .tr_interior_point import tr_interior_point
 
-
-__all__ = ['equality_constrained_sqp']
+__all__ = ['equality_constrained_sqp',
+           'tr_interior_point']
