+++ scipy/optimize/_large_scale_constrained/equality_constrained_sqp.py
