+++ scipy/optimize/_large_scale_constrained/projections.py
