+++ scipy/optimize/_large_scale_constrained/qp_subproblem.py
