+++ scipy/optimize/_large_scale_constrained/setup.py
@@ -3,7 +3,7 @@ from __future__ import division, print_function, absolute_import
 
 def configuration(parent_package='', top_path=None):
     from numpy.distutils.misc_util import Configuration
-    config = Configuration('_equality_constrained_sqp', parent_package, top_path)
+    config = Configuration('_large_scale_constrained', parent_package, top_path)
     config.add_data_dir('tests')
     return config
