+++ scipy/optimize/_large_scale_constrained/tests/test_projections.py
