+++ scipy/optimize/_large_scale_constrained/tests/test_qp_subproblem.py
